pyproject.toml
[project]
name = "product_description_agent"
version = "1.0.0"
description = "Product Description agent for the A2A - ADK Ecommerce Media Agency Project."
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.5",
    "google-adk>=1.2.1",
    "python-dotenv",
    "uvicorn",
    "python-dotenv>=1.1.0"
] 
